+++ src/GlobalPayments.Api/Gateways/GpApiConnector.cs
@@ -159,7 +159,7 @@ namespace GlobalPayments.Api.Gateways {
         public GpApiTokenResponse GetAccessToken() {
             AccessToken = null;
 
-            Request request = GpApiSessionInfo.SignIn(GpApiConfig.AppId, GpApiConfig.AppKey, GpApiConfig.SecondsToExpire, GpApiConfig.IntervalToExpire, GpApiConfig.Permissions);
+            Request request = GpApiConfig.AccessTokenProvider.SignIn(GpApiConfig.AppId, GpApiConfig.AppKey, GpApiConfig.SecondsToExpire, GpApiConfig.IntervalToExpire, GpApiConfig.Permissions);
 
             string response = base.DoTransaction(request.Verb, request.Endpoint, request.RequestBody);